requirements.txt
mysqlclient
django
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
django-cors-headers